ArchiveCenter 생성: 아카이빙/언아카이빙 기능 담당
- AppDelegate / SceneDelegate의 아카이빙 관련 기능 이동
