+++ VendingMachineApp/VendingMachineApp/AppDelegate.swift
@@ -11,48 +11,15 @@ import UIKit
 class AppDelegate: UIResponder, UIApplicationDelegate {
     
     var vendingMachine: VendingMachine!
+    private let archivingCenter = ArchivingCenter()
     
     func application(_ application: UIApplication, didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]?) -> Bool {
         
-        loadAll()
+        self.vendingMachine = archivingCenter.loadVendingMachine()
         return true
     }
     
-    enum Keys {
-        static let vendingMachine = "vendingMachine"
-    }
-
-    private func loadAll() {
-        if let vendingMachineData = UserDefaults.standard.data(forKey: Keys.vendingMachine),
-           let vendingMachine = unarchive(with: vendingMachineData) as? VendingMachine {
-            self.vendingMachine = vendingMachine
-        } else {
-            self.vendingMachine = VendingMachine(storage: BeverageStorage(),
-                                                 dispensedList: DispensedList(),
-                                                 moneyBox: MoneyBox(),
-                                                 beverageManager: BeverageManager(temperatureStandard: 36.5,
-                                                                                  sugarStandard: 1.0,
-                                                                                  lactoStandard: 3.5))
-        }
-    }
-
-    private func unarchive(with text: Data) -> Any? {
-        do {
-            let object = try NSKeyedUnarchiver.unarchiveTopLevelObjectWithData(text)
-            return object
-        } catch {
-            print(error)
-        }
-        return nil
-    }
-    
-    func archive(with things: Any) -> Data {
-        do {
-            let archived = try NSKeyedArchiver.archivedData(withRootObject: things, requiringSecureCoding: false)
-            return archived
-        } catch {
-            print(error)
-        }
-        return Data()
+    func save() {
+        archivingCenter.save(vendingMachine)
     }
 }
